+++ resources/views/livewire/like-article.blade.php
@@ -1,19 +1,14 @@
 <div class="relative">
-    @if ($showLikers)
-        <x-tooltip>
-            {{ $likers }} liked this article
-        </x-tooltip>
-    @endif
     @guest
         <div class="flex items-center text-gray-900 gap-x-4 gap-y-2.5 {{ $isSidebar ? 'flex-col' : 'flex-row' }}">
-            <x-heroicon-o-hand-thumb-up class="{{ $isSidebar ? 'w-6 h-6' : 'w-8 h-8' }}" wire:mouseover="toggleLikers" wire:mouseout="toggleLikers" />
+            <livewire:like :likable="$this->article" :isSidebar="$isSidebar" />
             <span class="font-medium leading-none {{ $isSidebar ? 'text-base' : 'text-2xl' }}">
                 {{ count($this->article->likes()) }}
             </span>
         </div>
     @else
         <button class="flex items-center text-gray-900 gap-x-4 gap-y-2.5 {{ $isSidebar ? 'flex-col' : 'flex-row' }} {{ $article->isLikedBy(Auth::user()) ? 'text-lio-500' : 'text-gray-900' }}" tag="button" wire:click="toggleLike">
-            <x-heroicon-o-hand-thumb-up class="{{ $isSidebar ? 'w-6 h-6' : 'w-8 h-8' }}" wire:mouseover="toggleLikers" wire:mouseout="toggleLikers" />
+            <livewire:like :likable="$this->article" :isSidebar="$isSidebar" />
             <span class="font-medium leading-none {{ $isSidebar ? 'text-base' : 'text-2xl' }}">
                 {{ count($this->article->likes()) }}
             </span>
